Added command tree syncing
The slash command tree is now copied to the guild whenever the bot runs.
That way, when slash commands are added in the future, they'll be
automatically added to the given guild.

This needs to be made much neater, but I wanted to add what I have
working so far.

Thanks to the discord.py Discord community for helping me out here!

Related to #13