Fixed sync command waiting
In the past, running the bot too many times too quickly would present a
message saying "We are being rate limited." This makes the bot run
without needing to wait for the commands to sync. It's not the
prettiest, but it works well enough for now until an owner sync command
is implemented.