Subtract the number of guilds by one
The `- 1` is needed since when adding the bot to a guild, it says the
bot is in one less guild than it actually is in (probably since it
doesn't count the bot's support server towards the count). The -1
corrects this to make the number seem more accurate (though it really is
now inaccurate).