Added check for proper id range
`get_user_id_from_string` would previously return improper ids like 32
if they were passed as parameters. This is no longer the case. The
method now ensures that all ids are within the proper range. All ids are
17 or more characters long
https://www.reddit.com/r/discordapp/comments/a5wtl4/lowest_id_on_discord
and 18446744073709551615 = `(2^64)-1` is the maximum possible id length
(https://discord.com/developers/docs/reference#snowflakes). My maximum
is a bit overkill, but this will ensure that I won't need to update this
in the future.