Quick bug fix
If a user was paused in a server, the server's admin couldn't change them to be the current user (maybe a user isn't able to be on consistently but the team wants to let them write every so often). This will let paused users be set as the current user.